requirements.txt
colorama
cycler
fonttools
joblib
kiwisolver
matplotlib
numpy
opencv-python
packaging
Pillow
pyparsing
python-dateutil
scikit-learn
scipy
six
threadpoolctl
torch
torchvision
tqdm
typing_extensions
